+++ src/App.css
@@ -52,15 +52,6 @@ body {
   align-items: center;
 }
 
-.todo-list-section {
-  margin-top: 25px;
-  border-width: 2px;
-  border-color: aliceblue;
-  padding: 10px;
-  border-radius: 10px;
-  background-color: aliceblue;
-}
-
 .todo-text-section {
   display: flex;
   flex-direction: row;
@@ -71,13 +62,15 @@ body {
 
 .todo-title, .done-title {
   padding:10px;
-  background-color: brown;
+  background-color: rgb(207, 207, 207);
+  color:black;
   width: 130px;
 }
 
 .todo-content, .done-content {
   padding:10px;
-  background-color: brown;
+  background-color: rgb(225, 225, 225);
+  color:black;
   width: 550px;
 }
 
@@ -91,10 +84,5 @@ body {
 
 .delete-btn {
   width: 120px;
-  background-color: rgb(79, 0, 0);
-}
-
-.complete-btn {
-  width: 120px;
-  background-color: rgb(0, 79, 46);
+  background-color: rgb(151, 57, 57);
 }
