+++ src/WorkingSection.css
@@ -1,0 +1,13 @@
+.working-list-section {
+    margin-top: 25px;
+    border-width: 2px;
+    border-color: aliceblue;
+    padding: 10px;
+    border-radius: 10px;
+    background-color: aliceblue;
+}
+
+.complete-btn {
+    width: 120px;
+    background-color: rgb(30, 93, 66);
+}
